requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
aiosmtplib==1.1.6
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Bootstrap-Flask==2.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
dnspython==2.2.1
easyforms==0.1.24
email-validator==1.1.3
Flask==2.1.0
Flask-Admin==1.6.0
flask-allows==0.7.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.0
Flask-Discussion==0.1.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Mailing==0.2.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
flask-oidc==1.4.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==1.0.0
fonttools==4.31.2
frozenlist==1.3.0
geoip2==4.5.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
idna==3.3
IPy==1.1
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
littlefish==0.0.70
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
maxminddb==2.2.0
multidict==6.0.2
numpy==1.22.3
oauth2client==4.1.3
packaging==21.3
pandas==1.4.1
passlib==1.7.4
Pillow==9.0.1
pony==0.7.16
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
PyMarkovChain==1.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
speaklater==1.3
SQLAlchemy==1.4.32
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.0.0
WTForms==3.0.1
yarl==1.7.2

weasyprint~=55.0
pdfkit~=1.0.0